refactor(form): Give each multi-select component its own file
Changes:
- Split MultiSelectGroup and MultiSelectSeparator out of the item file
- Extract the badge overflow measuring into use-badge-overflow
- Default-export every component and name them back up in the barrel

The family mixed one default export with five named ones and kept three
components in a single file, which is what AGENTS.md asks not to do. The
barrel keeps re-exporting under the old names, so no call site changes.

Every file is now inside the 50 to 100 line band except two eleven-line
passthroughs.

Notes:
- No behaviour change. Kept separate from the previous commit so a rename never hides one.
- Review finding 26 (reviews/REVIEW-2026-08-02-MULTI-SELECT-MOBILE-BY-AREA.md).
